Adding Vending Machine app
A Java program which stocks a vending machine using an input text file, receives user  input from the console, records dollar bill  inputs, purchases, and change returns to a log file  with the date and time, and produces a sales report when the machine is shut off. The program  has JUnit tests, and a Javadoc for the main class.